# openapi.yaml
openapi: 3.1.0
info:
  title: ReviveForge API
  version: 1.0.0
  description: |
    ReviveForge is a marketplace for discovering, valuing, and relaunching abandoned digital projects. This API powers all backend operations, including project listings, offers, payments, AI valuation, and user management.
servers:
  - url: https://{VERCEL_URL}/api
    variables:
      VERCEL_URL:
        default: your-vercel-deployment-url
security:
  - supabaseAuth: []
  - stripeSignature: []
paths:
  /projects:
    get:
      tags: [Projects]
      summary: List and filter projects
      parameters:
        - in: query
          name: techStack
          schema: { type: array, items: { type: string } }
        - in: query
          name: minPrice
          schema: { type: integer }
        - in: query
          name: maxPrice
          schema: { type: integer }
        - in: query
          name: minHealth
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  count:
                    type: integer
    post:
      tags: [Projects]
      summary: Create a new project
      security:
        - supabaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
  /stripe/webhooks:
    post:
      tags: [Payments]
      summary: Stripe webhook endpoint
      security:
        - stripeSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
...
  /user/profile:
    get:
      tags: [User]
      summary: Get user profile
      security:
        - supabaseAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
    put:
      tags: [User]
      summary: Update user profile
      security:
        - supabaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
...
components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT Bearer token
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: Stripe webhook signature header
  schemas:
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        seller_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        github_url: { type: string }
        tech_stack: { type: array, items: { type: string } }
        health_score: { type: integer }
        value_min: { type: integer }
        value_max: { type: integer }
        ai_analysis: { $ref: '#/components/schemas/AiAnalysis' }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectCreate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        github_url: { type: string }
        tech_stack: { type: array, items: { type: string } }
        health_score: { type: integer }
        value_min: { type: integer }
        value_max: { type: integer }
        status: { type: string }
...
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        stripe_account_id: { type: string }
        user_type: { type: string, enum: [buyer, seller] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProfileUpdate:
      type: object
      properties:
        name: { type: string }
        bio: { type: string }
        website: { type: string }
...
    AiAnalysis:
      type: object
      properties:
        healthScore: { type: integer }
        valueMin: { type: integer }
        valueMax: { type: integer }
        difficultyRating: { type: integer }
        analysisSummary: { type: string }